cookcli: Update to v0.8.0
I've actually never built a formula before but I found that this was
stuck on v0.1.6, so I wanted to take a whack at it. I used the
[cookbook] (hehe) to generate a new formula based on the tarball, so it
also renamed this to match the repo name. That's easy enough to undo but
I've included a [rename] entry for that as well.

[cookbook]: https://docs.brew.sh/Formula-Cookbook
[rename]: https://docs.brew.sh/Rename-A-Formula

This version will build the UI and CLI directly, and obviously uses NPM
and Cargo rather than Swift.

cookcli: rename from cook